# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Money API",
        "contact": {}
    },
    "paths": {
...
        "/transactions/entries/{period}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List a detailed view of entries joined with transactions for a given period",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List entries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "period in format YYYYMM",
                        "name": "period",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "name",
                        "description": "Sort field",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "Sort order (asc/desc)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of entries",
                        "schema": {
                            "$ref": "#/definitions/transactions.ListEntriesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/transactions/income": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create an income transactions and entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Create an income",
                "parameters": [
                    {
                        "description": "Income payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transactions.CreateIncomeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Income transaction created",
                        "schema": {
                            "$ref": "#/definitions/transactions.CreateIncomeResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/transactions/installment": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create an installment transactions and entries",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Create an installment",
                "parameters": [
                    {
                        "description": "Installment payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transactions.CreateInstallmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Transaction created",
                        "schema": {
                            "$ref": "#/definitions/transactions.CreateInstallmentResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/transactions/simple-expense": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a Simple Expense transaction and entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Create a Simple Expense",
                "parameters": [
                    {
                        "description": "Simple Expense payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transactions.CreateSimpleExpenseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Simple Expense created",
                        "schema": {
                            "$ref": "#/definitions/transactions.CreateSimpleExpenseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/transactions/{transaction_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a transaction and all entries related by the ID of the transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Delete Transaction By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transaction ID",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Transaction deleted"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "transactions.CreateIncomeRequest": {
            "type": "object",
            "required": [
                "amount",
                "name",
                "reference_date"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "maximum": 999999,
                    "minimum": 0
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "reference_date": {
                    "type": "string"
                }
            }
        },
        "transactions.CreateIncomeResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/transactions.CreateIncomeResponseData"
                }
            }
        },
        "transactions.CreateIncomeResponseData": {
            "type": "object",
            "properties": {
                "entry": {
                    "$ref": "#/definitions/transactions.ViewEntry"
                }
            }
        },
        "transactions.CreateInstallmentRequest": {
            "type": "object",
            "required": [
                "name",
                "reference_date",
                "total_amount",
                "total_installments"
            ],
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "reference_date": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number",
                    "maximum": 999999,
                    "minimum": 0
                },
                "total_installments": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                }
            }
        },
        "transactions.CreateInstallmentResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/transactions.CreateInstallmentResponseData"
                }
            }
        },
        "transactions.CreateInstallmentResponseData": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transactions.ViewEntry"
                    }
                }
            }
        },
        "transactions.CreateSimpleExpenseRequest": {
            "type": "object",
            "required": [
                "amount",
                "name",
                "reference_date"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "maximum": 999999,
                    "minimum": 0
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "reference_date": {
                    "type": "string"
                }
            }
        },
        "transactions.CreateSimpleExpenseResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/transactions.CreateSimpleExpenseResponseData"
                }
            }
        },
        "transactions.CreateSimpleExpenseResponseData": {
            "type": "object",
            "properties": {
                "entry": {
                    "$ref": "#/definitions/transactions.ViewEntry"
                }
            }
        },
        "transactions.ListEntriesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/transactions.ListEntriesResponseData"
                },
                "query": {
                    "$ref": "#/definitions/utils.QueryMeta"
                }
            }
        },
        "transactions.ListEntriesResponseData": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transactions.ViewEntry"
                    }
                }
            }
        },
        "transactions.TransactionType": {
            "type": "string",
            "enum": [
                "simple_expense",
                "income",
                "installment"
            ],
            "x-enum-varnames": [
                "SimpleExpense",
                "Income",
                "Installment"
            ]
        },
        "transactions.ViewEntry": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category_color": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "installment": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "period": {
                    "type": "string"
                },
                "reference_date": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "total_installments": {
                    "type": "integer"
                },
                "transaction_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/transactions.TransactionType"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "utils.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/utils.HTTPErrorData"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "utils.HTTPErrorData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "utils.QueryMeta": {
            "type": "object",
            "properties": {
                "next_page": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}